refactor request for quilt list
