+++ app/src/actions/quilts.js
@@ -1,0 +1,24 @@
+import {
+  REQUEST_QUILTS,
+  RECEIVE_QUILTS,
+} from '../constants/ActionTypes';
+import ip from '../config';
+
+const requestQuilts = () => ({
+  type: REQUEST_QUILTS,
+});
+
+const receiveQuilts = (quilts) => ({
+  type: RECEIVE_QUILTS,
+  payload: quilts,
+});
+
+export function fetchQuilts(token) {
+  return (dispatch) => {
+    dispatch(requestQuilts());
+    return fetch(`http://${ip}:8000/api/quilt?token=${token}`)
+      .then(response => response.json())
+      .then(data => dispatch(receiveQuilts(data)))
+      .catch(error => console.error('Error in getting user\'s quilts', error));
+  };
+}
